requirements.txt
boto3>=1.34.0
fastapi>=0.109.0
uvicorn>=0.27.0
gradio>=4.0.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-multipart>=0.0.6
pillow-heif>=0.18.0

# POC Dashboard
streamlit>=1.40.0
python-dotenv>=1.0.0

# Vision Providers
google-cloud-vision>=3.7.0
openai>=1.50.0
